Backend Guide
This is the backend guide I created back in February, but previously only showed it on Discord. Now it's on GitHub, with the PDF file with the step-by-step changes made, and I added the files that I adjusted in that zip file as well. Reach out to me if any issues occur. I will post a new guide today(3/14) as well.