Added Cart DB Functionality
Added DB functionality for the cart. It basically captures a snapshot of the logged in users cart contents and saves it to the DB.

If the user removes an item or decreases the quantity of an item, the changes will be reflected in the DB. Same goes for when a user completely removes an item from the cart.

For guest users, the cart fallsback to localStorage. Otherwise the cart feature wouldn't have worked at for guests.